refactor: replace nock with local server
