+++ lib/core/http.js
@@ -9,7 +9,6 @@ const phin = require('phin');
  * @returns
  */
 module.exports = (appUuid, apiKey, config) => {
-  console.log('HTTP', config);
   const request = (
     method,
     path,
