+++ lib/index.js
@@ -10,7 +10,6 @@ const {
   httpsHelper,
   attest,
 } = require('./utils');
-console.log('process', process.env.EV_API_URL);
 const config = require('./config');
 const {
   Crypto,
@@ -67,7 +66,6 @@ class EvervaultClient {
         Buffer.from(apiKey.slice(3), 'base64')
       );
     }
-
     this.config = config;
     let curve;
     if (!options.curve || !this.config.encryption[options.curve]) {
