+++ package.json
@@ -40,6 +40,7 @@
     "asn1js": "^3.0.5",
     "async-retry": "^1.3.3",
     "crc-32": "^1.2.2",
+    "env-cmd": "^10.1.0",
     "phin": "^3.5.0",
     "uuid": "^8.1.0"
   },
